+++ src/op_analytics/datapipeline/models/compute/auxtemplate.py
